jobflow-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobflow-mcp"
version = "0.1.0"
description = "Claude MCP 서버 — 개인 태스크 관리 도구"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "cryptography>=42.0",
    "pygithub>=2.1",
    "httpx>=0.27",
    "click>=8.1",
    "python-dateutil>=2.9",
]

[project.scripts]
jobflow = "jobflow_mcp.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src/jobflow_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
